chore: elixir update; necessary updates to make it compile, build and run
