+++ .tool-versions
@@ -1,4 +1,4 @@
-erlang 27.3.2
-elixir 1.18.3
+erlang 28.1.1
+elixir 1.19.3
 nodejs 20.11.0
 helm 3.18.3
