+++ lib/data_aggregator/application.ex
@@ -27,10 +27,10 @@ defmodule DataAggregator.Application do
       DataAggregator.PubSub,
       # Start the Finch HTTP client for sending emails
       {Finch, name: DataAggregator.Finch},
-      # Start the Oban queue
-      {Oban, Application.fetch_env!(:data_aggregator, Oban)},
       # Start the AshAuthentication system
       {AshAuthentication.Supervisor, otp_app: :data_aggregator},
+      # Start the Oban queue
+      {Oban, Application.fetch_env!(:data_aggregator, Oban)},
       # Start the Endpoint (http/https)
       DataAggregatorWeb.Endpoint
     ]
