+++ lib/data_aggregator/files/workers/sdf.json
@@ -1,0 +1,34 @@
+{
+  "results": [
+    {
+      "__metadata": {
+        "id": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='E0003')",
+        "uri": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='E0003')",
+        "type": "ZCO_MASTERDATA_API_SRV.ZCO_R_WBS_STATUSType"
+      },
+      "WBSElementExt": "B-ZN369003.E.0001.0007",
+      "Status": "E0003",
+      "StatusShortName": "PLEP"
+    },
+    {
+      "__metadata": {
+        "id": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='I0002')",
+        "uri": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='I0002')",
+        "type": "ZCO_MASTERDATA_API_SRV.ZCO_R_WBS_STATUSType"
+      },
+      "WBSElementExt": "B-ZN369003.E.0001.0007",
+      "Status": "I0002",
+      "StatusShortName": "REL"
+    },
+    {
+      "__metadata": {
+        "id": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='I0460')",
+        "uri": "https://CHSU0049.CE.INT.AMECSPIE.COM:443/zext_odata/ZCO_MASTERDATA_API_SRV/ZCO_R_WBS_STATUS(WBSElementExt='B-ZN369003.E.0001.0007',Status='I0460')",
+        "type": "ZCO_MASTERDATA_API_SRV.ZCO_R_WBS_STATUSType"
+      },
+      "WBSElementExt": "B-ZN369003.E.0001.0007",
+      "Status": "I0460",
+      "StatusShortName": "ACPT"
+    }
+  ]
+}
